Initial commit
I hope I've got roughly the right things included/not included but I'm
far from confident. I had to discuss this with ChatGPT as Android Studio
doesn't seem to be willing to offer an opinionated default.

This is really me playing around anyway so not too critical. This commit
does have a "working as far as I've implemented it" combo box, so it's
good to have this to return to if I mess it up with future changes.