Tidy layouts, pseudo-break animation
The debug Red colour is not generally visible, but shows up dramatically
during the animation between Home and Settings. This may be a valuable
clue so I haven't tried reverting it yet. I think the general top-level
Compose structure of both screens is now more regular (which may be
neater and/or help animation), and we are also no longer incorrectly
applying the screenBorder padding around the TopAppBar.