Possibly tweaked/fixed/better animations
I can't find any clear guidance on the MD3 "standards" for transition
animations, but ChatGPT was reasonably convincing about these options
(while also saying I could just use EaseInOur for symmetry, but
suggesting this was perhaps better). I think this does look OK, albeit
too slow - but I want the slow animations for now so I can keep an eye
on them and watch for glitches. A bit of superficial testing on the P7
suggests that these are consistent, it may be that the weird
inconsistencies will start to reappear with a more realistic 300 tween
duration but let's stick with this for now.