Broken in an interesting way
My home screen is now broken with the keyboard scroll - a key part of
the problem is having the "current thing" far enough down the screen. I
don't know why this is broken, but it suggests that Dialog may *not*
have actually been the culprit, and that with further hair-pulling this
can be fixed. Since I now have a reproduction with no Dialog in sight,
it *ought* to be possible to make this work, which is heartening. When I
had a Dialog involved, the (possibly incorrect) advice I was getting
from AIs (because I couldn't find any other advice) was that Dialog and
on-screen keyboards didn't get on, and so it felt like there might not
*be* a solution, but now it feels like there will be. This is so
fundamental it must work, right? But probably yes, right.