Fix for white space issue
We needed to move the imePadding() right on to the very top level
Surface() inside our theme. With this both the main screen and the
screen-pretending-to-be-a-dialog work. It's quite possible some of the
other changes added are not necessary, haven't tried to investigate that
yet. And of course, this tweak may fix things for the Dialog approach.