Hacky conversion to Flow instead of LiveData
This is probably what I should have all along, but I got missteered by
trying to ask an AI to generate me some simple code instead of just
diving in with the Rooms-based tutorial that uses Flow.

I was/am trying to avoid getting a concrete database in because it's
just going to be an extra level of "oh, now the database has changed and
I can't reproduce that" complexity right now.