Possible bug fix
State updates get coalesced and if the state changes A->B->A in a single
frame, that does *not* count as a change and does not trigger e.g.
LaunchedEffect(). ChatGPT says to use an "event" to avoid this, but we
still need the state as the event is just a "yes, something happened"
type signal. This was already hellish and gets more so by the minute.