Revert unintended change to unit choice for unit price on main screen
I may well add a drop down here, but the recent unit family rework broke
this. The demo milk data showed the problem. We are trying to show (as
default, if we add a dropdown) the unit price in the ItemSource card
using the same unit the user entered the price in. So if the user
entered the price using pints, we should unit the relevant unit family
including pints for the unit price, not any unit family which the user
has enabled for volume. (Luckily things worked out so the code chose
litres as the best unit making the problem noticeable, had it happened
to choose pint anyway this wouldn't have been visible.)