Possibly working variable unit price denominator
The code is a bit untidy but this does seem to work and I suspect (can't
test until I actually add price editing) it is robust and efficient-ish.
I may be deluding myself but it feels like I actually understand what's
going on (mostly) with the recomposition and remember stuff.