Snapshot with debug code
So I was logging dataSetListRaw not oldUIState.dataSetList(Raw - this is
just a naming quirk in this hacky code) and that was confusing me into
thinking rememberSaveable() was not preserving my state so it was
instantly available on the first frame after going back to this screen
from another, when it was working fine all along. A couple of hours
wasted^Winvested in education.

Had some interesting and probably incorrect discussions with AIs, hence
the saveableStateHolder stuff which does not (based on the todoWTF
experimental loggic) appear to actually be necessary - even without it,
rememberSaveable() appears to be preserved when we come back from the
settings screen or the edit screen.