Move code following ChatGPT example, fix crash
The basic move here was based on ChatGPT's refactor of this code, which
I asked for after I'd done mine. (It got it subtly wrong the first time,
hah.) That then triggered a crash trying to do
getPreference(SELECT_DATA_SET_ID_KEY), so I've fixed a possible
misnaming (application -> app) in the voodoo line, and moved the
definition of app up to the top as I think that may affect
initialisation order. This seems to work. This voodoo around preference
context worries me, but sufficient unto the day.