Avoid re-querying all data sets every time
Lightly tested and seems to work. I don't have an easy way to add new
data sets to check they are going to be picked up by the flow, but it
should be fine. And it should also be fairly obvious if this doesn't
work once I add that "add data set" functionality too.