Possibly efficient and consistent querying!
We still don't have any "loading" timeouts, but one step at a time. As
far as I can tell from watching logcat output, this is only delivering
consistent state to the UI and is only querying the database (within the
context of the home screen anyway; it's possible there's an inefficiency
with the edit dialog, but that code needs reworking anyway) as
necessary.