"Fix" CurrencyTextField
I had incorrectly changed this from how Grok wrote it. The problem is
that because numberFormat is currency-based, it *expects* the string to
be a valid currency including symbol.

I think the code as written is extremely silly - you can delete the €
symbol and that breaks the parsing - but I will commit this and maybe
tweak further, as it still feels like there might be the core of a good
idea here.