Hack up CurrencyTextField
This has the pseudo-cleverness I (at least thought) I perceived in the
original sort of working - as you type, grouping separators are
introduced, the currency symbol can't be deleted, etc, because we
round trip the string via double and formatCurrency() each time it's
edited. However, this doesn't really work as the cursor sort of moves
around when you type but it isn't always responding to what you type
(e.g. you can type letters on hardware keyboard and they move the
cursor).

Gut feeling is this approach isn't great, but I wanted to try it. The
hints on how to format currency etc are at least helpful. I suspect I
may end up trying to format the number as currency, infer the
prefix/suffix characters then apply those to a textfield which is purely
number and maybe (either myself or via some standard support) try to
restrict the textfield to numeric input in a less weird-feeling way.

Previous commit was broken, newT should have been newText but not going
to bother to fix it.