Remove NumericTextFieldValue
The double-remember hack is probably kind of workable anyway. As
probably already noted, a proper use of this control would probably
store the TextFieldValue in the ViewModel, although I suspect like
everything else this will be hellishly complex (and that EditablePrice
is actually mostly useless) because everything has to be a state-based
thing so the GUI updates when values update, but that's not specific to
this.

We don't need the hackery in lost focus here - we probably wouldn't
anyway now the TextFieldValue itself is not saved across rotations, but
I suspect even when it moves into ViewModel it will be fine. My best
guess - and ignoring why rememberSaveable() didn't seem to actually
be getting a chance to discard the non-text state, although relying on
that would probably be bad UI anyway - is that TextField internally
updates the value you pass it (not unreasonably - it *is* a
TextFieldState) when things like losing focus happen, but with our
maintaining a NumericTextFieldState externally that was *not* updated on
losing focus, when we recomposed after a rotation our
NumericTextFieldState still had the selection in. (It always did, but it
wasn't being used normally, as there was no recompose, I suspect.) I may
not be making sense there but it's the best I can do.

I only tried adding NumericTextFieldValue in order to avoid clients (i.e.
me) from trying to use rememberSaveable with the TextFieldValue and it
seeming to work until they happen to test a rotation. By making the API
use NTFV and making that parcelable, I figured it would "just work" if a
client did do that - but as per waffling in paragraph above the
situation is more complex and it feels better just to use the raw
TextFieldValue. It's not as if Compose isn't littered with other
landmines where things compile and then some innocuous action blows
things up anyway.