Get rid of randomUUID on navigation route
Testig, this seems to work fine. I think the way we deliberately
re-create the EditPriceScreenViewModel in the context of our current
back stack entry ensures that we don't get re-use. FWIW I had a chat
with Grok about this and this is all probably reasonable design given
how my app works, in particular that this is an "edit" flow where we
don't want stale data hanging around.