Remove nullability inside EditPriceViewModel
This is where the refactor commits were heading. It is maybe not such a
big win as I'd hoped, but barring lurking fundamental problems with this
approach it does feel cleaning. It's also kind of nice that the recovery
from a SavedStateHandle happens in the same place as "normal"
construction, rather than being buried inside EditPriceViewModel.