WIP simplification of data handling for general selector
This seems to be working for the store selector version. The idea here -
as I hinted at yesterday in my comments about SavedStateHandle - is that
if the app is killed and resurrected, we don't really need any saved
state (beyond the data set ID and name, which we can pass as route
parameters and which are automatically resurrected for us) as it's
just a micro-optimisation for jank-free animation between screens, and
we don't care about that when we're being resurrected.