Remove TODOs after discussing composable identity
ChatGPT and Grok have both clarified (and this does feel right) that
composable *identity* is based on position in the compose hierarchy.
Changes in the parameters to a child composable cause it to be
re-composed, but *not* to changes its identity. So the behaviour I was
observing with the just-deleted log line where the textFieldWidth *was*
being remembered across enabled changing from true to false is
explained. (I liked the behaviour - textFieldWidth specifically probably
didn't matter anyway, but it was proving a point - but previously didn't
see why this desirable behaviour was occurring.)