Fix minor bug, add all the boilerplate for edit data sets
The copy and paste here is awful but I will do the data set dialog and
then see what commonalities can be factored out - it feels risky to do
this before there are actually two "similar things".

This also fixes a probably-unimportant copy and paste bug in the "set
shared view model copy to null" for "editSource".