Half-decent effort at dropdown currency selector
If we ignore the possible dropdown last item visibility bug and the
missing divider, this isn't great but it's not terrible either. For an
app not primarily concerned with currencies, I think this is easily good
enough even for a first full release, even though there's obviously
scope for further polish later on.