"Fix" safeDrawingPadding and nav drawer size
It looks like we don't want or need the very top level
safeDrawingPadding - everything "just works" anyway (probably the
Scaffold stuff is applying it internally) and now the navigation drawer
also has what I believe is the standard behaviour of going to the top
and bottom of the screen but the relevant areas still work to navigate
in system-controlled ways.

I do wonder if this would also have fixed whatever problem I was having
where I just could not get things to quite line up - probably when I was
trying to get the home screen loading scrim to cover *just* the part
below the combo boxes at the top. But it doesn't really matter about
that, it *was* fiddly anyway and I already decided it's such a rare case
that the chances of it breaking on some random phone and hardly getting
any testing are too high to be worth it..