WIP factoring out
The saving isn't huge and it may be we never save enough lines of code
to compensate for the generic framework, but as per one of the new TODOs
I suspect there is more simplification to be had, and this does feel
somewhat cleaner. I mostly understand the new code, except for the
noinline @DisallowComposableCalls and crossinline
annotations/qualifiers/whatever the hell they are.