Icon tweaks
I suspect changing the rectangular paths to use mitre joins doesn't harm
(and may improve slightly) the appearance, while retaining a round join
for the pencil avoids the ugly appearance of a mitre (or less ugly but
not great) bevel join at the tip. This will reduce the number of ugly
and/or faffy curves then the icon is traced.