Keep per-quantity type default measure unit during editing
FFS^FFS. Also there *may* have been a lurking parcelize bug only visible
when doing the kill-and-revive dance with the old code, since it would
probably try to serialize "defaultUnit: MeasureUnit". I am not 100% sure
about this and right now I have not tried reverting to the old code to
test. OK, I just tried it and it didn't seem to suffer from this. It may
be there is a lurking bug somewhere around this in the new code then,
although I haven't been able to trigger it and it may have been a
"development app upgrade" occurrence of the Android shooting me in the
foot problem I ranted about in one of the new TODO comments.