Tweak home screen spacing
I think this does generally look better - it means the card contents
align with the text in the two TextFields at the top. It is squeezing my
Price as sold/Confirmed/Unit price layout but that maybe already needed
tweaking anyway. (This would just get worse if I did an outer screen
border of 16.dp at left/right instead of 8.dp as well.)