More icon tweaks
Not necessarily great but this doesn't feel too bad.

I didn't like using the warning triangle in not red, and it was also
confusing to re-use it for "stale data" when we use it more
conventionally for validation failures. And the red error circle felt a
little over-dramatic even for ancient prices.