Add currently unused locale to price history delta generation
This is a precursor to filtering out no-visible-diff changes, which I
need the locale for as we need to format the confirmed at date as a
date only. Actually I half wonder if I'd have got away with this
regardless, because it's *timezone* that matters for this, not locale.
But it is probably still cleaner for the history to be pure strings and
do the diffs on those.