Tweak TODO after failed attempt to implement it
Trying to inline this breaks as we're no longer invoking
dropUnlessResumed() from a @Composable, and it isn't possible to just
annotate the lambda as @Composable. I may be missing something there. I
talked to ChatGPT and Grok and the latter offered a mildly clever and
acceptable way to pull this sort of thing out into a reusable modifier,
but I don't need that anywhere else (at least yet) and it was verbose
and slightly over-clever, and all the other workarounds were worse than
what we have here.