Remove TODO
The times are just milliseconds since epoch going into
getRelativeTimeSpanString, so there are no timezone considerations at
that point. I think it will be fine, certainly with respect to the
problem I mentioned in the TODO, and if there are any very odd corner
cases we can worry about them later or perhaps just accept them - we
don't necessarily want to be getting involved in complex timezone stuff
given the nature of this app.