Rework RelativeTimeText()
This did in practice seem to work but it may have been a bit unclear
(the dependence on state variable now being updated to trigger
recompositions wasn't explicit) and it may also have had subtle bugs. I
think this new version is clearer, assuming it is in fact correct.
Lightly tested and seems to be working.

The use of System.currentTimeMillis() was *not* wrong, but it feels a
lot clearer and slightly more consistent to be using "now" instead.