Try to fix quirks with "per each"
Not sure this is perfect but it's probably better. I am trying not to
get too hung up on translation issues, I hope this approach will in
fact be generic but if it's not I can make future code tweaks once the
actual translation requirements are clear because I'm speaking with a
translator.