Tweak comments
I suspect the SavedStateHandle comment is true, but it's far from
exhaustive - I think we'd also need to write to SSH in onValueChange
too, and maybe other places. Seems a bit silly to complicate the code
with half-arsed comments for a future enhancement which I ought to be
able to think through at the time I make it, if I do.