Make restoreDatabase() check version before overwriting
We need a temp file so we can check the version on the file the user has
asked us to import before we overwrite our internal database. Admittedly
if the user wants to import something, the chances are our internal data
isn't valuable, but still better to avoid blatting it and also probably
crashing by overwriting it with something we know we can't handle.

This code is based on Grok's from the restore-hell branch and a
discussion with ChatGPT.