Change how STALE_PRICE_THRESHOLD_KEY is defaulted
It's probably cleaner to just default to the default (so to speak)
rather than explicitly handling null everywhere. What I'm really
thinking about here is in future when a new version adds a new
preference - we want to be consistent how we handle preferences, and
it's probably cleaner to accept the presence of a null and treat it as a
default rather than trying to set the preference to a default when the
app is upgraded.