Tweak layout of Settings screen
I was just using ChatGPT's code before and the left margins were not the
same as my e.g. "Edit Product" screen using GeneralSelectorListItem and
they didn't left-align with the back arrow in the scaffold top bar. This
is still probably wrong - it's MD3, so getting things right is
impossible - but at least it's consistently wrong and the left alignment
of the arrow and items are approximately the same.