Tweak spacing
As always this feels like a complete nightmare and it still feels a bit
of a mess, but it's probably better than it was. It does feel as if the
"left edge" of the screen content should more-or-less align with the
left edge of the top bar back arrow, regardless of MD3's half-arsed
prescriptions, and I think this mostly does that.