Experimental rewrite of loading timeout on home screen
Superficially works, very lightly tested, but I've been hacking around
for ages and I really want to get something comitted. Lots of dead code
from earlier attempts in here, semi-deliberately so it can be referred
to if I need to come back to this and try out some variant ideas.