Preserve modifiedAt on prices for confirm/undo confirm
This seems to work, and is of course actually simpler. I suspect this
was just a consequence of perhaps a) feeling that it *ought* to change
before (as per TODO I deleted a few commits ago), which I now disagree
with b) rushing to bash some code out when I first wanted to implement
this feature.