Tweak compileSdk, set version to 0.1
compileSdk probably doesn't matter much but I think it ought to match
targetSdk.

I want the initial version to be <1.0 when released, and there's also a
vague reference to Acorn OS 0.1.

I think for now I won't show the internal version code in the about
screen, it just confuses things and it's not like we have lots of
internal build versions floating around with a big team etc.