Minor tidying
I am not sure the "Maybe we should attach this modifier to the
ErrorHighlightBox" TODO is actually meaningful given the current state
of the code. I can't really see what I meant, given that the one place
this modifier is used *is* inside the implementation of
ErrorHighlightBox. I may just have been bashing out a random thought or
this may have code-rotted as things evolved, but I don't think it's
helpful right now.