Rename PriceTrackerRepository to Repository
The app is probably not going to be called "Price Tracker", and I don't
want to (and am not sure I should) name the repository after the app
anyway. I only have one repository so I am not giving it a "what it
contains" type name (which is probably the "rule" I misinterpreted as
suggesting the app name should be used), albeit PriceRepository wouldn't
be terrible.