Minor tidying
The coroutine scope leaks were false alarms; there is a *general* risk
with this sort of thing, but here - where we are writing to the global
application context-tied datastore - it is fine. I think.

We still have unpleasantly and probably unavoidably duplicated code, but
we have at least simplified the viewmodel functions a bit.