Tidy settings read/writes
This is slightly more verbose and (annoyingly) we *still* have to refer
to the default value for each setting (albeit via a constant, as we
already did) twice, but I think on the whole it is clearer and the
verbosity is not extreme.

AppScope.io is a Grok suggestion. It feels reasonably sound but I am
obviously no expert. ChatGPT was lecturing me about not using
CoroutineScope(Dispatchers.IO) and telling me I ought to have callers
pass scopes in, despite having previously told me it was fine for what
I'm doing here (updating a global DataStore). I suspect using a passed
in scope might actually be technically incorrect - if they were
ViewModelScopes or rememberCoroutineScope()s and the screen went away
quickly enough, the coroutine might cancel before the job runs, if I
understand Grok's apparently sensible reasoning correctly.