Minor tidying
I think the "layout here is good" TODO is just badly out of date. It
doesn't seem to match the code and as far as I can see
runGeneralEditScreenOperation() is updating asyncOperationStatus nicely.

The buffer() comment tweaks are maybe a tiny bit of a cop out but I
think it's probably a sound general principle to have this buffer(), and
in practice it doesn't seem to hurt even if it may not be strictly
necessary.