Tweak validation messages
Grok is very keen on "Required" and even if it's not as widely used as
it makes out it does feel broadly reasonable here. I also don't want to
have to write custom "must have a price", "must have a pack size"
messages analogous to the former "must have a name" validation for each
numeric validation cases.