Minor tidying
Plausible discussion with ChatGPT suggests neither of these is an issue.
I am already creating and re-using the formatter objects themselves
outside the LazyColumn. We don't need a key because the default Compose
behaviour of identifying items by index is absolutely fine here - the
list is immutable within any once instance of this screen.