Move price edit quantity to its own row
I was really reluctant to do this but even with English having all three
things on the same row was just a tiny bit tight on small phone and even
on medium phone Spanish was still not great. Let's see how I feel
about this as I get used to it. Right now it doesn't seem too bad,
albeit a little bit of a shame.