Convert SelectItem* to PersistentUiContent, tweak GeneralSelectorViewModel
To make this work as cleanly and safely as possible, I've stopped
GeneralSelectorViewModel actually being a ViewModel, and that means I
had to actually create a trivial SelectDataSetViewModel wrapper around
it instead of using it directly. SelectSource* (and if it's necessary
but it may not be SelectDataSet*) have not been converted to
PersistentUiContent, the changes to them are just to keep up with the
other changes.

Lightly tested and seems to work.