Upgrade dependencies
I don't specifically need any of these but Android Studio was warning me
and since I haven't done that much "final testing" yet it seemed a good
idea to do this now. Inevitably some hair-pulling with build failures
during this but it does build now and the app superficially appears to
work fine.