Remove TODOs, tweak naming
Let's stick with "toDomain()" - it's probably technically correct, it
does make some kind of sense, it will avoid renaming everything. I've
renamed toHistory because it feels like it ought to use the target class
name - this is an Entity<->Entity (although also Domain) conversion,
which is a different kind of thing anyway from toDomain().