Minor tidying
Decide not to use EditXScreenStaticContent() in the data set and item
cases, as they have no or a single item or static data and it would just
be foolish consistency - the data set is already a bit different anyway.
I think using EditSourceScreenStaticContent does mildly improve the code
so I have done that.