Remove leading capitals from Spanish demo data
This is consistent with keyboard capitalization, and based on chatting
with both ChatGPT and Perplexity it is what a native speaker would
expect. The lack of a capital when one of these values is used in the
top app bar as a title is apparently absolutely expected by a native
speaker, even if it looks odd to my English eyes. Any kind of
auto-capitalisation on the top app bar title would be inappropriate
(which is a relief as I'm sure it would turn into a mess of corner
cases).