Minor tidying
On the price-vs-value terminology, ChatGPT and Perplexity both (probably
unbiasedly prompted) seemed to lean towards "price" (value has overtones
of judging quality etc as well) and I was semi-leaning that way myself.
I don't think it's a huge deal but this acts as a sort of confirming
vote, and I need to not obsess over this sort of stuff too.