Minor tidying
diff() is a symmetric function, it might be OK for it to be an extension
on lhs or rhs but it's not by any means critical, and it is used in
precisely one place by other code in this file and it's private. I am
waffling here, but in general I am trying to just stop obsessing over
tiny details and get all these goddamn TODOs sorted. ;-)