More log tidying
It feels slightly destructive to rip all this out, but the reality is
most of this is stuff I wrote while I was feeling my way during
development. It isn't obvious most of this would have any major
support/debugging value, and of course I can easily put it back in later
if I need it (and probably do it better, in the context of whatever I am
investigating).