Rename UserPreferences message to CurrentSelections
Massive boilerplate change, because reasons. But lightly tested and it
does seem to work, and this probably saves pain in the long run if I
want to add another message and don't want one of them to be
irrationally called UserPreferences when all of my hypothetical messages
relate to some aspect of user preferences.