Use ValidateFieldState in SettingsDialog
I think this is a bit cleaner, and more to the point it gives us the
same delayed supportingText appearance we have in our other dialogs for
improved consistency.

As part of doing this I've also tweaked the handling of empty text
fields in SettingsDialog - we now show no message (but the save button
remains greyed out, as it always did) in this case.