Add some imePadding modifiers
This is slight voodoo but it does seem to help with e.g. the loyalty
discount disappearing under the on screen keyboard (on Android 10
at least) and the user not being able to drag to scroll the screen to
see it properly, and giving it a quick test it doesn't appear to break
anything.

The behaviour is still not great at least on Android 10 but attempting
to futz around trying to call bringIntoView() when the TextField gets
focus was not working well and I gave up. This is still an improvement
and is probably harmless even if it doesn't help in all cases.